:truck: Separated EMA and unconditional checkpoint tests
